+++ src/core/server/saved_objects/import/create_saved_objects.ts
@@ -39,9 +39,9 @@ interface CreateSavedObjectsParams<T> {
   importIdMap: Map<string, { id?: string; omitOriginId?: boolean }>;
   namespace?: string;
   overwrite?: boolean;
-  workspaces?: string[];
   dataSourceId?: string;
   dataSourceTitle?: string;
+  workspaces?: string[];
 }
 interface CreateSavedObjectsResult<T> {
   createdObjects: Array<CreatedObject<T>>;
@@ -59,9 +59,9 @@ export const createSavedObjects = async <T>({
   importIdMap,
   namespace,
   overwrite,
-  workspaces,
   dataSourceId,
   dataSourceTitle,
+  workspaces,
 }: CreateSavedObjectsParams<T>): Promise<CreateSavedObjectsResult<T>> => {
   // filter out any objects that resulted in errors
   const errorSet = accumulatedErrors.reduce(