+++ src/core/server/saved_objects/import/import_saved_objects.ts
@@ -55,9 +55,9 @@ export async function importSavedObjectsFromStream({
   savedObjectsClient,
   typeRegistry,
   namespace,
-  workspaces,
   dataSourceId,
   dataSourceTitle,
+  workspaces,
 }: SavedObjectsImportOptions): Promise<SavedObjectsImportResponse> {
   let errorAccumulator: SavedObjectsImportError[] = [];
   const supportedTypes = typeRegistry.getImportableAndExportableTypes().map((type) => type.name);