+++ src/core/server/saved_objects/import/resolve_import_errors.ts
@@ -59,9 +59,9 @@ export async function resolveSavedObjectsImportErrors({
   typeRegistry,
   namespace,
   createNewCopies,
-  workspaces,
   dataSourceId,
   dataSourceTitle,
+  workspaces,
 }: SavedObjectsResolveImportErrorsOptions): Promise<SavedObjectsImportResponse> {
   // throw a BadRequest error if we see invalid retries
   validateRetries(retries);
@@ -163,9 +163,9 @@ export async function resolveSavedObjectsImportErrors({
       importIdMap,
       namespace,
       overwrite,
-      workspaces,
       dataSourceId,
       dataSourceTitle,
+      workspaces,
     };
     const { createdObjects, errors: bulkCreateErrors } = await createSavedObjects(
       createSavedObjectsParams