+++ src/core/server/saved_objects/import/types.ts
@@ -187,10 +187,10 @@ export interface SavedObjectsImportOptions {
   namespace?: string;
   /** If true, will create new copies of import objects, each with a random `id` and undefined `originId`. */
   createNewCopies: boolean;
-  /** if specified, will import in given workspaces */
-  workspaces?: string[];
   dataSourceId?: string;
   dataSourceTitle?: string;
+  /** if specified, will import in given workspaces */
+  workspaces?: string[];
 }
 
 /**
@@ -212,10 +212,10 @@ export interface SavedObjectsResolveImportErrorsOptions {
   namespace?: string;
   /** If true, will create new copies of import objects, each with a random `id` and undefined `originId`. */
   createNewCopies: boolean;
-  /** if specified, will import in given workspaces */
-  workspaces?: string[];
   dataSourceId?: string;
   dataSourceTitle?: string;
+  /** if specified, will import in given workspaces */
+  workspaces?: string[];
 }
 
 export type CreatedObject<T> = SavedObject<T> & { destinationId?: string };