+++ src/core/server/saved_objects/service/saved_objects_client.ts
@@ -71,6 +71,10 @@ export interface SavedObjectsCreateOptions extends SavedObjectsBaseOptions {
   initialNamespaces?: string[];
   /** permission control describe by ACL object */
   permissions?: Permissions;
+  /**
+   * workspaces the new created objects belong to
+   */
+  workspaces?: string[];
 }
 
 /**