+++ src/plugins/workspace/server/saved_objects/saved_objects_wrapper_for_check_workspace_conflict.test.ts
@@ -33,6 +33,9 @@ describe('WorkspaceConflictSavedObjectsClientWrapper', () => {
   };
   wrapperInstance.setSerializer(savedObjectsSerializer);
   describe('createWithWorkspaceConflictCheck', () => {
+    beforeEach(() => {
+      mockedClient.create.mockClear();
+    });
     it(`Should reserve the workspace params when overwrite with empty workspaces`, async () => {
       mockedClient.get.mockResolvedValueOnce(
         getSavedObject({
@@ -83,6 +86,34 @@ describe('WorkspaceConflictSavedObjectsClientWrapper', () => {
           }
         )
       ).rejects.toThrowError('Saved object [dashboard/dashboard:foo] conflict');
+    });
+
+    it(`Should use options.workspaces when get throws error`, async () => {
+      mockedClient.get.mockRejectedValueOnce({
+        output: {
+          statusCode: 404,
+        },
+      });
+
+      await wrapperClient.create(
+        'dashboard',
+        {
+          name: 'foo',
+        },
+        {
+          id: 'dashboard:foo',
+          overwrite: true,
+          workspaces: ['bar'],
+        }
+      );
+
+      expect(mockedClient.create).toBeCalledWith(
+        expect.anything(),
+        expect.anything(),
+        expect.objectContaining({
+          workspaces: ['bar'],
+        })
+      );
     });
   });
 
@@ -135,6 +166,10 @@ describe('WorkspaceConflictSavedObjectsClientWrapper', () => {
           getSavedObject({
             id: 'bar',
             workspaces: ['foo', 'bar'],
+          }),
+          getSavedObject({
+            id: 'qux',
+            workspaces: ['foo'],
           }),
         ],
       });
@@ -199,6 +234,7 @@ describe('WorkspaceConflictSavedObjectsClientWrapper', () => {
             id: 'qux',
             references: [],
             type: 'dashboard',
+            workspaces: ['foo'],
           },
         ],
         {
@@ -240,6 +276,15 @@ describe('WorkspaceConflictSavedObjectsClientWrapper', () => {
               "workspaces": Array [
                 "foo",
                 "bar",
+              ],
+            },
+            Object {
+              "attributes": Object {},
+              "id": "qux",
+              "references": Array [],
+              "type": "dashboard",
+              "workspaces": Array [
+                "foo",
               ],
             },
           ],